refactor: consolidate run module
